race fetch bug fix
Implemented backup API, called when a race is just finished and so the results are not available yet. The corresponding card in PastEvents Activity will show a reminder on the temporarily  unavailability of the results.
TO FIX: when attempting to enter the event page of the just-finished race, an error is encountered in LocalRaceResult, not being able to fetch race results from local database, causing the app to crash.